# Backend/labfastorder-openapi.yaml
openapi: 3.1.0
info:
  title: "labfastorder API"
  description: "labfastorder API"
  version: "1.0.0"
servers:
  - url: "localhost:8080"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderDetailDTO:
      type: "object"
      properties:
        orderDetailId:
          type: "integer"
          format: "int64"
        quantity:
          type: "integer"
          format: "int32"
        price:
          type: "number"
          format: "double"
        productId:
          type: "integer"
          format: "int64"
        productName:
          type: "string"
        productSize:
          type: "string"
...
    OrderDTO:
      type: "object"
      properties:
        orderId:
          type: "integer"
          format: "int64"
        status:
          type: "string"
        datetime:
          type: "string"
        totalPrice:
          type: "number"
          format: "double"
        orderDetails:
          type: "array"
          items:
            $ref: "#/components/schemas/OrderDetailDTO"
        tableId:
          type: "integer"
          format: "int64"
...
    LoginRequest:
      type: "object"
      properties:
        username:
          type: "string"
        password:
          type: "string"
    SignupRequest:
      type: "object"
      properties:
        username:
          type: "string"
        email:
          type: "string"
        role:
          type: "array"
          items:
            type: "string"
        password:
          type: "string"
    JwtResponse:
      type: "object"
      properties:
        token:
          type: "string"
        type:
          type: "string"
          default: "Bearer"
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        email:
          type: "string"
        roles:
          type: "array"
          items:
            type: "string"
    MessageResponse:
      type: "object"
      properties:
        message:
          type: "string"
...
security:
  - bearerAuth: []
paths:
...
  /api/orderDetails:
    get:
      summary: "GET api/orderDetails"
      operationId: "getAllOrderDetails"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/OrderDetailDTO"
  /api/orderDetails/{id}:
    get:
      summary: "GET api/orderDetails/{id}"
      operationId: "getOrderDetailById"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDetailDTO"
...
  /api/orders:
    get:
      summary: "GET api/orders"
      operationId: "getAllOrders"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/OrderDTO"
    post:
      summary: "POST api/orders"
      operationId: "createOrder"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                orderDetails:
                  type: "array"
                  items:
                    properties:
                      quantity:
                        type: "integer"
                        format: "int32"
                      product:
                        type: "object"
                        properties:
                          productId:
                            type: "integer"
                            format: "int64"
                tableId:
                  type: "integer"
                  format: "int64"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /api/orders/open:
    get:
      summary: "GET api/orders/open"
      operationId: "getAllOpenOrders"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/OrderDTO"
  /api/orders/{id}:
    get:
      summary: "GET api/orders/{id}"
      operationId: "getOrderById"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    delete:
      summary: "DELETE api/orders/{id}"
      operationId: "deleteOrder"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
  /api/orders/{id}/status:
    patch:
      summary: "PATCH api/orders/{id}/status"
      operationId: "updateOrderStatus"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              properties:
                status:
                  type: "string"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
...
  /api/auth/signin:
    post:
      summary: "POST api/auth/signin"
      operationId: "authenticateUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtResponse"
  /api/auth/signup:
    post:
      summary: "POST api/auth/signup"
      operationId: "registerUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
...